chore(lib/ethclient): wire headerdb
